+++ components/shared/index.js
@@ -29,3 +29,4 @@ export {
 export { WeekGrid } from "./WeekGrid";
 export { Timeline } from "./Timeline";
 export { PaceRunner } from "./PaceRunner";
+export { DayAgenda } from "./DayAgenda";